:truck: Split API endpoint tests from forms app into separate test modules
